policy added to operations

+++ app/Http/Controllers/ApplicationController.php
@@ -11,6 +11,7 @@ use App\Models\Application;
 use App\Models\Resume;
 use Illuminate\Support\Facades\Auth;
 use App\Models\User;
+use App\Models\JobPost;
 
 
 class ApplicationController extends Controller
@@ -21,6 +22,9 @@ class ApplicationController extends Controller
     }
     public function index()
     {
+        if (Auth::user()->cannot('viewAny',Application::class)) {
+            return redirect()->route('home')->with('error', 'sorry but you do not have the privilage to do this operation.');
+        } 
         $applications = Application::where('status', '!=', 'cancelled')->paginate(6);
         return view('application.index', compact('applications'));
         
@@ -28,6 +32,10 @@ class ApplicationController extends Controller
     
     public function getMyApplications(User $user)
     {
+        if (Auth::user()->cannot('viewMyPostsApplications', $user)) {
+            return redirect()->route('home')->with('error', 'sorry but you do not have the privilage to do this operation.');
+        } 
+
         $applications = Application::where('status', '!=', 'cancelled' )->where('user_id', $user->id)->paginate(6);
         return view('application.index', compact('applications'));
         
@@ -36,6 +44,14 @@ class ApplicationController extends Controller
     {
         $user = Auth::user();
         $resumes = Resume::where('user_id', $user->id)->get();
+        $jobPost = JobPost::findOrFail($job_id);
+        if (Application::where('user_id' , Auth::id())->where('job_id' , $job_id)->exists()) {
+            
+            $application = Application::where('user_id' , Auth::id())->where('job_id' , $job_id)->first();
+            // dd($application);
+            return redirect()->route('applications.edit' , $application->id)->with('error', 'sorry but you do not have the privilage to do this operation.');
+        } 
+
 
         return view('application.add_application', compact('resumes', 'user', 'job_id'));
     }
@@ -43,8 +59,18 @@ class ApplicationController extends Controller
     public function store(StoreApplicationRequest $request)
     {
         $validatedData = $request->validated();
+
+        // $jobPost = JobPost::findOrFail($validatedData['job_id']);
+
+        // if (Auth::user()->cannot('create',$jobPost)) {
+            
+        //     $application = Application::where('user_id' , Auth::id())->first();
+        //     // dd($application);
+        //     return redirect()->route('applications.edit' , $application->id)->with('error', 'sorry but you do not have the privilage to do this operation.');
+        // } 
+
         $validatedData['user_id'] = Auth::user()->id;
-        $validatedData['status'] = 'pending'; // Default status
+        $validatedData['status'] = 'pending'; 
 
         if ($request->hasFile('resume')) {
             $resumePath = $request->file('resume')->store('CVs', 'uploaded_files');
@@ -94,7 +120,9 @@ class ApplicationController extends Controller
 
     public function show(Application $application)
     {
-
+        if (Auth::user()->cannot('view', $application)) {
+            return redirect()->route('home')->with('error', 'sorry but you do not have the right to do this operation.');
+        }
         $resume = Resume::findOrFail($application->resume_id);
  
         return view('application.view_application',['application'=> $application , 'resume'=> $resume] );
@@ -103,6 +131,9 @@ class ApplicationController extends Controller
     public function updateStatus(Request $request, Application $application)
     {
 
+        if (Auth::user()->cannot('approveOrCancel', $application)) {
+            return redirect()->route('home')->with('error', 'sorry but you do not have the right to do this operation.');
+        }
 
         $validatedData = $request->validate([
             'status' => 'required|in:pending,approved,cancelled',
@@ -115,6 +146,10 @@ class ApplicationController extends Controller
     }
 
     public function destroy(Application $application){
+        if (Auth::user()->cannot('delete', $application)) {
+            return redirect()->route('home')->with('error', 'sorry but you do not have the right to do this operation.');
+
+        } 
         $application->delete();
         return back()->with('success', 'application deleted successfully Deleted successfully!');
 
@@ -122,6 +157,12 @@ class ApplicationController extends Controller
 
     
     public function cancel(Application $application){
+
+
+ if (Auth::user()->cannot('approveOrCancel', $application)) {
+            return redirect()->route('home')->with('error', 'sorry but you do not have the right to do this operation.');
+
+        } 
         $application->status='cancelled';
         $application->save();
